Organize zetachain repo (#115)
