+++ packages/addresses/src/addresses.athens.json
@@ -8,6 +8,7 @@
     "multiChainSwap": "0x8BD7144Ddb59c9Fa3Dcf809998521E9cAD946fa1",
     "multiChainSwapZetaConnector": "",
     "multiChainValue": "0x064516547ECd3b2D1709e1b2798Aae92b1C8a84C",
+    "rewardDistributorFactory": "",
     "tridentPoolFactory": "",
     "tss": "0x8531a5aB847ff5B22D855633C25ED1DA3255247e",
     "tssUpdater": "0x55122f7590164Ac222504436943FAB17B62F5d7d",
@@ -33,6 +34,7 @@
     "multiChainSwap": "0x323745f16C93e56a98012970c28788498d8B3a14",
     "multiChainSwapZetaConnector": "",
     "multiChainValue": "0x14BeC0E4A8e7bF7A02Af54Ad81a57a9fcA4D37Fd",
+    "rewardDistributorFactory": "",
     "tridentPoolFactory": "",
     "tss": "0x8531a5aB847ff5B22D855633C25ED1DA3255247e",
     "tssUpdater": "0x55122f7590164Ac222504436943FAB17B62F5d7d",
@@ -56,6 +58,7 @@
     "dai": "",
     "immutableCreate2Factory": "0x095a03c6a68137fE9a566bBc3e552F299d8b886d",
     "multiChainSwap": "0xb1b812b664c28E1bA1d35De925Ae88b7Bc7cdCF5",
+    "rewardDistributorFactory": "",
     "tridentPoolFactory": "",
     "multiChainSwapZetaConnector": "",
     "multiChainValue": "0x14BeC0E4A8e7bF7A02Af54Ad81a57a9fcA4D37Fd",
@@ -82,6 +85,7 @@
     "immutableCreate2Factory": "",
     "multiChainSwap": "",
     "multiChainValue": "",
+    "rewardDistributorFactory": "",
     "tridentPoolFactory": "",
     "tss": "",
     "tssUpdater": "",
@@ -107,6 +111,7 @@
     "multiChainSwap": "",
     "multiChainSwapZetaConnector": "",
     "multiChainValue": "0x82aC45D07dEe4DBDe050e838beF345347DEd99a8",
+    "rewardDistributorFactory": "0x3a969d797A28d61e8127562944E639a5824F1219",
     "tridentPoolFactory": "",
     "tss": "",
     "tssUpdater": "",
@@ -117,8 +122,8 @@
     "uniswapV3Router": "",
     "usdc": "",
     "weth9": "0x5F0b1a82749cb4E2278EC87F8BF6B618dC71a8bf",
-    "zetaSwap": "0xd886b7Af031F9a505310bA01951948BD1d673aF1",
-    "zetaSwapBtcInbound": "0x064516547ECd3b2D1709e1b2798Aae92b1C8a84C",
+    "zetaSwap": "0x44D1F1f9289DBA1Cf5824bd667184cEBE020aA1c",
+    "zetaSwapBtcInbound": "0x008b393933D5CA2457Df570CA5D628380FFf6da4",
     "zetaToken": "0x5F0b1a82749cb4E2278EC87F8BF6B618dC71a8bf",
     "zetaTokenConsumerUniV2": "",
     "zetaTokenConsumerUniV3": ""
@@ -132,6 +137,7 @@
     "multiChainSwap": "",
     "multiChainSwapZetaConnector": "",
     "multiChainValue": "",
+    "rewardDistributorFactory": "",
     "tridentPoolFactory": "",
     "tss": "",
     "tssUpdater": "",
