+++ packages/zevm-example-contracts/package.json
@@ -33,7 +33,7 @@
     "@uniswap/v2-periphery": "1.1.0-beta.0",
     "@zetachain/addresses": "workspace:^",
     "@zetachain/addresses-tools": "workspace:^",
-    "@zetachain/protocol-contracts": "^2.0.1",
+    "@zetachain/protocol-contracts": "^2.1.0",
     "ethers": "5.6.8"
   }
 }
